Replaced .stream().parallel() with .parallelStream(), now /omcs refresh runs in another thread and sends more feedback, replaced SupplierThrows with Callable

+++ src/main/java/net/zatrit/skins/SkinsCommands.java
@@ -4,8 +4,8 @@ import com.moandjiezana.toml.Toml;
 import com.mojang.brigadier.CommandDispatcher;
 import com.mojang.brigadier.context.CommandContext;
 import dev.isxander.yacl3.config.v2.api.ConfigClassHandler;
-import lombok.AllArgsConstructor;
 import lombok.Cleanup;
+import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
 import lombok.val;
 import net.fabricmc.fabric.api.client.command.v2.ClientCommandRegistrationCallback;
@@ -19,19 +19,23 @@ import net.zatrit.skins.config.SkinsConfig;
 import net.zatrit.skins.util.command.*;
 import org.apache.commons.io.FileUtils;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.util.concurrent.CompletableFuture;
 
 import static com.mojang.brigadier.arguments.IntegerArgumentType.integer;
+import static net.zatrit.skins.lib.util.SneakyLambda.sneaky;
 import static net.zatrit.skins.util.command.CommandUtil.argument;
 import static net.zatrit.skins.util.command.CommandUtil.literal;
 
-@AllArgsConstructor
+@RequiredArgsConstructor
 public class SkinsCommands implements ClientCommandRegistrationCallback {
     private final ConfigClassHandler<SkinsConfig> configHolder;
     private final HasAssetPath assetPath;
+    private @Nullable CompletableFuture<Void> cleanupFuture;
 
     @Override
     public void register(
@@ -44,8 +48,7 @@ public class SkinsCommands implements ClientCommandRegistrationCallback {
                 new IndexedResourceProvider(
                         "presets",
                         getClass().getClassLoader()
-                ),
-                new DirectoryFileProvider(presetsPath)
+                ), new DirectoryFileProvider(presetsPath)
         }, "toml");
         presetsType.refresh();
 
@@ -65,18 +68,11 @@ public class SkinsCommands implements ClientCommandRegistrationCallback {
                 // omcs list
                 .then(literal("list").executes(this::listHosts))
                 // omcs remove (id)
-                .then(literal("remove").then(argument(
-                        "id",
-                        integer(0)
-                ).executes(this::removeHost)))
+                .then(literal("remove").then(argument("id", integer(0)).executes(
+                        this::removeHost)))
                 // omcs move (from) (to)
-                .then(literal("move").then(argument(
-                        "from",
-                        integer(0)
-                ).then(argument(
-                        "to",
-                        integer(0)
-                ).executes(this::moveHost))));
+                .then(literal("move").then(argument("from", integer(0)).then(
+                        argument("to", integer(0)).executes(this::moveHost))));
 
         dispatcher.register(command);
         dispatcher.register(literal("omcs").redirect(command.build()));
@@ -186,14 +182,34 @@ public class SkinsCommands implements ClientCommandRegistrationCallback {
     @SuppressWarnings("resource")
     private int clean(
             @NotNull CommandContext<FabricClientCommandSource> context) {
-        Files.list(Path.of(assetPath.getAssetPath()).resolve("skins"))
-                .map(Path::toFile).parallel().forEach(directory -> {
-                    try {
-                        FileUtils.deleteDirectory(directory);
-                    } catch (IOException e) {
-                        SkinsClient.getErrorHandler().accept(e);
-                    }
-                });
+        if (cleanupFuture != null && !cleanupFuture.isDone()) {
+            context.getSource().sendError(Text.translatable(
+                    "openmcskins.command.cleanupAlready"));
+            return -1;
+        }
+
+        cleanupFuture = CompletableFuture.<Void>supplyAsync(sneaky(() -> {
+            Files.list(Path.of(assetPath.getAssetPath()).resolve("skins")).map(
+                    Path::toFile).parallel().forEach(directory -> {
+                try {
+                    FileUtils.deleteDirectory(directory);
+                } catch (IOException e) {
+                    SkinsClient.getErrorHandler().accept(e);
+                }
+            });
+
+            return null;
+        })).whenComplete((r, e) -> {
+            if (e == null) {
+                context.getSource().sendFeedback(Text.translatable(
+                        "openmcskins.command.cleanupSuccess"));
+            } else {
+                context.getSource().sendError(Text.translatable(
+                        "openmcskins.command.cleanupFailed",
+                        e.getMessage()
+                ));
+            }
+        });
 
         return 0;
     }
